Geetest-Solver/requirements.txt
opencv-python
rsa
pycryptodomex
curl_cffi